requirements.txt
#annotated-types
anyio
certifi
click
fastapi
h11
httpcore
httpx
idna
Pillow
pydantic
pydantic_core
pyngrok
python-dotenv
python-telegram-bot
PyYAML
sniffio
starlette
typing_extensions
uvicorn
psutil
